pyproject.toml
[project]
name = "recon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "click>=8.3.2",
    "colorlog>=6.10.1",
    "dnspython>=2.8.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "tabulate>=0.10.0",
]
